config data tests + main merged
